avoid circular import of _show_dataarray
