+++ plantcv/plantcv/plot_image.py
@@ -8,7 +8,7 @@ from plantcv.plantcv import params
 from plantcv.plantcv.classes import PSII_data
 from matplotlib import pyplot as plt
 from plantcv.plantcv import fatal_error
-from plantcv.plantcv._debug import _show_dataarray
+from plantcv.plantcv._show_dataarray import _show_dataarray
 
 
 def plot_image(img, cmap=None, **kwargs):
